+++ DifferentiationInterface/test/Core/Internals/context.jl
@@ -1,5 +1,5 @@
 using DifferentiationInterface
-using DifferentiationInterface: Rewrap, fix_tail
+using DifferentiationInterface: AnyConstant, GeneralizedConstant, Rewrap, fix_tail, unwrap
 using Test
 
 f1(x) = x
@@ -18,3 +18,17 @@ contexts = (Constant(1.0), Cache([2.0]))
 r = @inferred Rewrap(contexts...)
 @test (@inferred r(3.0, [4.0])) == (Constant(3.0), Cache([4.0]))
 @test (@inferred r(3, [4.0f0])) isa Tuple{Constant{Int}, Cache{Vector{Float32}}}
+
+contexts = (PrepTimeConstant(1.0), Cache([2.0]))
+r = @inferred Rewrap(contexts...)
+@test (@inferred r(3.0, [4.0])) == (PrepTimeConstant(3.0), Cache([4.0]))
+@test (@inferred r(3, [4.0f0])) isa Tuple{PrepTimeConstant{Int}, Cache{Vector{Float32}}}
+
+# PrepTimeConstant behaves like Constant in the type hierarchy
+@test PrepTimeConstant <: GeneralizedConstant
+@test PrepTimeConstant <: AnyConstant
+@test Constant <: AnyConstant
+@test !(Cache <: AnyConstant)
+@test !(ConstantOrCache <: AnyConstant)
+@test unwrap(PrepTimeConstant(2.0)) == 2.0
+@test PrepTimeConstant(2.0) == PrepTimeConstant(2.0)