+++ DifferentiationInterface/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/JuliaDiff/DifferentiationInterface.jl/compare/DifferentiationInterface-v0.7.21...main)
 
+### Added
+
+- Add the `PrepTimeConstant` context type, for constants whose value is fixed at preparation time and may therefore be specialized upon by backends ([#265](https://github.com/JuliaDiff/DifferentiationInterface.jl/issues/265))
+
 ## [0.7.21](https://github.com/JuliaDiff/DifferentiationInterface.jl/compare/DifferentiationInterface-v0.7.20...v0.7.21)
 
 ### Fixed